refactored stopwatch to use the system stopwatch function rather than unity's time.deltaTime

+++ Assets/Scripts/ScriptableVariables/RepMs.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [CreateAssetMenu]
 public class RepMs : ScriptableObject
 {
-    public float rep_ms;
+    public double rep_ms;
 }
